requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
langchain-openai
python-dotenv
opencv-python
mediapipe
numpy
sounddevice
vosk
pvrecorder
pvporcupine